[lib/network] Add test for or_terminal_event race
The race was described and fixed in https://fxbug.dev/328297563.

We pulled the Control::or_terminal_event() method into it's own
freestanding function to make it easier to test without having to drag
in all of the FIDL client code.  It's also unclear if we can reliably
trigger this bug if we tried to reproduce it at the FIDL level.

Ran against the previous code to make sure it fails reliably when the
bug is present.

Fixed: 329679745
Bug: 328297563
Change-Id: Ib9bdacb5edaa78f9d4611ccb8e8c6b08b8210b8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006854
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>