[fxfs] Refactor versioning scheme
The versioning scheme has been considerable source of confusion for Fxfs
developers.  This CL attempts to rectify that.

The common interpretation of FooV20 was that:
- Foo was changed at V20, and
- FooV20 represents the struct at that version.

This was inconsistent with the intended interpretation, which was that
FooV20 was the *last* version at which Foo looked like that struct, and
Foo was changed at some unknown prior version.

Based on discussions, this CL moves to the common interpretation; now
FooV20 means that Foo was changed at V20 to look like that struct.

I noticed while making this change that at least one version was done
incorrectly; new fields were added to ObjectKeyData in V33, but some of
the containing structs were not versioned.  This was corrected.

Fixed: 326988259
Change-Id: I7f01d3f6f33667847cb19f30af065d0116776b95
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007514
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>