[netstack3] Local ingress hook
Add the local ingress entry point for the filtering crate, which occurs
for incoming IP packets that have been routed for local delivery but
have not yet been dispatched to the transport layer.

This CL modifies the existing tests for the ingress hook to exercise the
local ingress hook as well, since, for the socket layer, there is no
observable difference between traffic being dropped at ingress or at
local ingress.

It also adds a separate flavor of integration test to exercise the
difference: in this test, instead of a client and server directly
communicating, we introduce a router in between them, and install
filters on the router. In this scenario, installing filters on local
ingress does nothing, because all client-server traffic is forwarded and
therefore skips the local ingress hook; installing filters on ingress,
however, does affect client-server connectivity.

Fixed: 326082742
Test: netstack-filter-fidl-test
Change-Id: Icaf095e01cf95291fdf1f0fd12d88e2d9d8c91d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005312
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>