[fidlc] Allow 0 as bits constant
This allows assigning the number zero to a bits constant, for example
`const EMPTY MyBits = 0;`. The number must be a literal, and it must be
zero. All other values can be expressed by ORing together bits members.
Zero was the exception to that, hence this change.

Test: fx test fidlc-test -- --gtest_filter="ConstsTests.GoodBitsTypedConstZero"
Test: fx check-goldens fidl && fx build //tools/fidl/fidlgen_{rust,hlcpp,cpp}:goldens
Fixed: 329872697
Bug: 42079239
Change-Id: Ic20a335ae8e8b791f2dd355466d94bd1312c290f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009444
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>