[wlan][policy] remove legacy file clean up
This change removes the code that was in place to delete the legacy file
being used for saved networks persistent storage. Before this, the
legacy file already wasn't being used if it had anything in it, just
deleted in case it was still around.

Bug: 42166225
Test: existing unit tests for loading saved networks verify that the rest of the behavior remains unchanged. This change just removes file cleanup that we don't need anymore.

Change-Id: I8d808c8d8db17aae8c1c0fa7615b97e7a9ce3bfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1008216
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>