[sdk] Adjust documentation for SAG's FIDL
Updates the documentation to indicate _how_ these protocols are expected
to be used.

Additionally, we remove the passive deps on WakeHandling and
ApplicationActivity because we do not see a use for them today.

Additionally, we convert the level constants to enums to ensure that the
level values do not overlap and to indicate that the values are related
to each other.

Change-Id: I98af7ecd077857bfceac744521500b2da9fd5f0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006813
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
API-Review: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>