[core_size_limit] Adjust the Wlan-Drivers creep_limit to a smaller one
Currently the creep_limit of Wlan-Drivers is larger than total blobfs
creep_limit, so that the limit will never be hit. This was because when
we introduce a new firmware to the driver, there'll be a big jump on the
binary size, and the huge creep_limit was to tolerant the size jump.
Given that adding a firmware is not a common case, decreasing the
creep_size to a reasonable value.

Change-Id: Ibfd559afa6e98a3c0881902e24d4bbd11965870f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010657
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>