[vfs][cpp] Move append mode handling to FileConnection
Move file-specific append state out of fs::Connection and into
fs::FileConnection, and refactor both RemoteFileConnection
and StreamFileConnection accordingly.

Remove options() from fs::Connection, and instead store only the rights
associated with the connection. This removes the need to filter specific
flags for querying or creating new connections, and allows handling them
only for the types of connections that the flags are associated with.

Update DirectoryConnection to return NOT_SUPPORTED when calling
SetFlags to match the behavior of the Rust VFS and SDK C++ VFS.

Bug: b/324112857
Test: fx test fs-connection-test
Change-Id: Iacc504833e1c8ab98509c3183f60d54c9ac4303e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007953
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>