[netfilter] Add port matching for fnet_filter
Add functionality for parsing port numbers into the types provided by
the new filter library. These functions are mostly the same as the
parser for the deprecated filter library. The same tests for the
deprecated parser have been mapped to the new filter library to ensure
the same functionality is supported for the same grammar.

Bug: 323949040
Change-Id: I5ed022d853962a31f3bb923b73c0f10eb470c967
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005647
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>