[logs] Add buffering on the ffx side for symbolizer writes
Buffer the symbolizer writes on the ffx side if the symbolizer
doesn't read from stdin in time and gets stuck.
This happens when a large volume of logs is sent to the symbolizer
at a faster rate than it can read from the buffer. We need to
read from the symbolizer's stdout concurrently in order to drain
it.

MULTIPLY: host_x64/ffx_log_lib_test

Bug: b/299980894
Change-Id: I93789d9668ea9711b74af3d760045d547baab534
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010935
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>