[time][test] Removes unneeded expose stanzas
After some swapping the context back in, it seems that these
exposes are a remnant of something I tried to do, which failed,
and I then never cleaned up the remnants of.

In turn, this means that `test.time.realm.RealmFactory` is
the only source of in-realm protocols that should be used.

Removing the stanzas and comment now, to reduce future
confusion potential.

Bug: b/327476961
Change-Id: Ic00a213af326f78568370ab1ed471ef42ca71790
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009326
Reviewed-by: Chris Phoenix <cphoenix@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>