[component_manager] Open takes DirectoryEntry impl instead of OpenFn
To convert an open capability into a VFS node, a single callback is
insufficient.  For the io2 migration, some of the callbacks supplied to
the open capability might need to support either the io1 or io2 open
request.  All of this is captured by the DirectoryEntry trait, so this
changes the open capability to accept something that implements the
DirectoryEntry trait.  Most of the changes here are to accommodate this
change.

This then revealed that the into_directory method on Open can't always
work; not everything can be represented as a directory.  To address the
use case for it, this CL adds a try_into_directory method on Capability
instead.  This replaces the try_into_open, into_directory combo.

There is one functional change here: the Sender/Receive capabilities now
terminate node connections on the Sender side.  It would probably make
sense to remove flags from these capabilities.

Change-Id: Icf4c4ad8ed2ffc0e100eeb92c935e1e4fa1779a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007132
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Rui Qi Sim <rqsim@google.com>