[component] Test lazy directory routing behavior
We currently wait for a directory request to be readable before handing
it off to the provider component. This way component-manager steps out
of the way on subsequent fs operations.

I didn't find a test of this behavior. I also didn't find a place
specifically for directory tests. There were some in routing testing
storage but not directory per se. So I added a new module.

Change-Id: I29bdc82127aa1ec61e798230e7bf5c5eb1621464
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011274
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Gary Bressler <geb@google.com>