[netstack3] Move TCP to new TimerContext
Introduce CoreTimerContext in the base crate, which provides the new
timer-transformation mechanism that ties in with the new timer context.

Use the fact that we now need a type to control timers to provide
type-safety around the fact that only connected TCP sockets have timers,
which is a nice side-effect.

While I'm here, we've identified a race with tw reuse that can happen
between the reuse and timewait state timers firing. Relax the timer
handler and timer cancelation to deal with that.

Bug: 42083407
Change-Id: I51faa7c6523860a63f468a68f2a1d6377f99e8e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009218
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zeling Feng <zeling@google.com>