[system-updater] Try the other slot when reading an asset errors
When checking to see if the non-current configuration already has the
correct asset fails with an error, system-updater used to fall back to
resolving the image package. Now it will first try to see if the
current configuration has the correct asset, the same as if checking
the non-current configuration failed just because the contents were
different.

Change-Id: I6fdc6e7efbe07c0f59b21af89d5b589829592767
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010814
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>