[bazel_sdk] Drop ABI with subpackages_to_flatten.
A subsequent change will forbid including the ABI stamp file in a
package manifest, since it could possibly conflict with the --api-level
specified on the CLI. This clears the way for that.

Also, adds a note to subpackages_to_flatten that it's a temporary
workaround.

Bug: 330189874
Bug: 326095332
Change-Id: If7599ed43ee29db5355d69bf12a9d619a98e60e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009256
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>