[ramdisk] Set dead_ on DdkAdd failures.
Failures to call DdkAdd currently shutdown by calling DdkRelease on the
device where it will wait for the worker thread to join. The worker
thread will only complete when dead_ is set which is done in DdkUnbind
which will not be called through the failure path for DdkAdd.

Also logs failures to DdkAdd to make it clear what the failure is.

Bug: b/326397893
Change-Id: I22752c70c34e21b2c35576c3cf645671bbbc68b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013634
Reviewed-by: Brandon Castellano <bcastell@google.com>
Fuchsia-Auto-Submit: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>