[aml-sdmmc] Minor refactoring in preparation for Power integration.
No behavioural changes are intended in this CL.

This change decouples power awareness from whether the incoming request
is of Banjo or FIDL. This is useful because Banjo and FIDL have
different request processing models. For example, this refactoring makes
it easy to delay a FIDL request and serve (complete) it at a later time
using the Async completer.

Change-Id: I3d390eb746b1d2fb4ec434dc222904b6d4ee8761
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011472
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>