[ffx] Rename to clarify role of target specifier
The name `get_default_target()` was misleading, as it didn't get
a target (it got a specifier) and it didn't get a default (e.g.
providing "-t <spec>" on the command line goes through the
same code).

To that end, the function, some variables, and some comments have
been updated to clarify the role of the function.

Bug: b/320519654
Change-Id: I57421925434bf4c130162689f98b3273a8e89b94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009226
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>