[fidl] Optimize the client waker
On Fuchsia, we can read the channel when we are woken and then only open
the task that needs to be woken rather than all of them.

On host, we can't do that because there are reentrancy issues, but we
can still avoid allocating a combined waker.

Change-Id: I4ff5dc6a8f96637e79b1fab1b7606fb0591aad60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010794
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>