[heapdump][instrumentation] Do not dynamically link libunwind.so.1
When libheapdump_instrumentation.so is used by SDK clients,
libunwind.so.1 may not exist or have a different ABI in the client's
toolchain.

We originally used to link it statically using the
//build/config/fuchsia:static_cpp_standard_library config option, but
fxrev.dev/919272 removed it to avoid a warning, introducing the runtime
dependency on libunwind.so.1 by mistake.

This CL restores the old correct behavior, but in a way that does not
trigger the warning.

Change-Id: Ia17e26285cd3f801fec1a1b8c283053e21f7496c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013656
Commit-Queue: Fabio D'Urso <fdurso@google.com>
Reviewed-by: Étienne J. Membrives <etiennej@google.com>