[fsverity] Increase the buf size for file reads on enable-verity
Enable verity reads the data attribute into memory in chunks while
computing the merkle tree. This CL increases the size of those chunks to
reduce the cost of block I/O. This size was chosen based on the size of
the largest verity-enabled file on boot. We can tune this further as we
get more data.

Bug: b/314836822
Change-Id: I9b4ffadb551b4cdd44439ec86c0fed67872cc076
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001974
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Nikita Jindal <nikitajindal@google.com>